Implement sparse * diagonal and diagonal * sparse operations, double-prec only.
Date: Sun, 8 Mar 2009 16:28:18 -0400
These preserve sparsity, so eye(5) * sprand (5, 5, .2) is *sparse*
and not dense.  This may affect people who use multiplication by
eye() rather than full().

The liboctave routines do *not* check if arguments are scalars in
disguise.  There is a type problem with checking at that level.  I
suspect we want diag * "sparse scalar" to stay diagonal, but we have
to return a sparse matrix at the liboctave.  Rather than worrying
about that in liboctave, we cope with it when binding to Octave and
return the correct higher-level type.

The implementation is in Sparse-diag-op-defs.h rather than
Sparse-op-defs.h to limit recompilation.  And the implementations
are templates rather than macros to produce better compiler errors
and debugging information.