Implement diag + sparse, diag - sparse, sparse + diag, sparse - diag.
Date: Mon, 9 Mar 2009 17:45:22 -0400
This does not use the typical sparse-mx-ops generator.  I suspect the
sematics of elementwise multiplication and division to be rather
controversial, so they are not included.  If comparison operations are
added, the implementation should be shifted over to use the typical
generator.

The template in Sparse-diag-op-defs.h likely could use const bools
rather than functional argument operations.  I haven't measured which
is optimized more effectively.

Also, the Octave binding layer in op-dm-scm.cc likely could use all
sorts of macro or template trickery, but it's far easier to let Emacs
handle it for now.  That would be worth revisiting if further
elementwise sparse and diagonal operations are added.