Do not reduce Magick::ErrorCoder to warning, issue an error like other errors.
* __magick_read__.cc (read_file): the Magick::ErrorCoder is issued by the
different coders and is being caught as a warning without any apparent reason.
A simple warning when no image was actually read leads to a segfault when
accessing the elements of imvec, which currently happens in F__magick_finfo__
An alternative would be to check the size of imvec, but really, an error here
should cause an error. GraphicsMagick documentation says that warnings are
for `suspected but completed' while errors are for `failed', so if anything
is issuing an error but still read the image and we should continue, then it
needs to be fixed upstream to throw a Magick::WarningCoder instead.

--HG--
branch : stable