Overhaul FLTK pan, rotate, zoom
* graphics.in.h: add axes properties pan, rotate3d, mouse_wheel_zoom
  and custom set_pan which disables rotate3d.

* graphics.cc: add custom set_rotate3d and link with pan property.
  Disable rotate3d for 2D plots.

* __init_fltk__.cc: replace gui_mode and mouse_wheel_zoom with axes
  properties pan, rotate3d and mouse_wheel_zoom. Disable pan for legends,
  move them instead.

* __add_default_menu__.m: Add new menu entries for new pan and zoom modes.

* findall.m: Update test for added uimenus.

Each axes now has its own properties for interactive GUI control of pan,
rotate3d and mouse_wheel_zoom. Now it's possible to have several figures
and set pan for the 2D plot in figure x and rotate3d for the 3D plot in
figure y. There are two new pan modes: "Pan x only" and "Pan y only".

The toolbar buttons "P" and "R" set pan and rotate3d for the last clicked axes
object or the object below the center of the canvas if none was clicked yet.

The legend can now be moved with the mouse.