Add "text/x-octave" and "text/x-matlab" as MIME types we are able to handle.
* etc/icons/octave.desktop.in: according to the latest desktop entry specs
(version 1.1), we can register the MIME types supported by this application.
These are "the MIME types that an application knows how to handle".
The only MIME types that freedesktop currently recognizes are "text/x-octave"
and "text/x-matlab". Once we officialy register a MIME type with IANA, we
should add a mime type for vendor .m octave files and possible .oct.
* etc/icons/Makefile.am: since we now define Octave supported MIME types,
we need to rebuild the MIME types application database as part of Octave
installation (unless we are not doing a desktop installation.