pkg install: drop INSTALLDIR and stop calling configure with --prefix.
* pkg/private/configure_make.m: do not export INSTALLDIR, and do not call
configure with "--prefix=INSTALLDIR".  This is not currently used.  It
can even be ignored in the case of binary packages.  Packages should be
robust to being built and then have their contents shuffled around later.
The recent changes to pkg/private/build.m also require this change (the
alternative fix would be to define desc.dir there but why would we?).