Rework color scaling for gnuplot toolkit (bug #48664)
* __gnuplot_draw_axes__.m: Remove 'cdatadirect' variable and operate directly
on individual object property 'cdatamapping'.  Remove 'truecolor' variable.
If axes cautoscale is set, do a first pass to get the 'clim' min and max of
'cdata' for all objects having 'scale' for 'cdatamapping' property.  If the
number of rows in the color map is not two, do a second pass to check for any
image data that is logical.  If logical image found, set axis parent color map
to black/white, axis 'clim' to [0 1] and 'climmode' to 'manual'.  Set 'cmap'
and 'cmap_sz' variables before the axes children loop.  If "image", "patch" or
"surface" object has 'cdatamapping' property set to scale, linearly scale
color data according to 'clim' using fix--not round--with range clamping.  The
one exception is "image" with logical 'cdata' in which case 1 is added to the
data (equivalent of 'clim' of [0 1]).  Simplify the logic for setting gnuplot
cbrange to [1,cmap_sz] or [1,cmap_sz+row(addedcmap)] for extended color maps.