doc: remove "Contributing guidelines" and "Tips and standards" from the manual.
* doc/interpreter/contrib.txi, doc/interpreter/tips.txi: remove sections
"Contributing guidelines" and "Tips and standards" which are outside of
the scope of a language manual.  The Octave website is a better fit for
this subject and most of this content has been merged in the Octave wiki
(most of it was already duplicated).
* doc/interpreter/external.txi: remove reference to removed section about
tips on writing good documentation.
* doc/interpreter/install.txi: replace link to removed section of the
manual with reference to the files in source that have instructions for
building.
* doc/interpreter/preface.txi: replace link to removed section of the manual
with link to octave.org for instructions on how to get involved.
* etc/HACKING: replace references to the removed sections of the manual
with link to the main page on the wiki with contribution guidelines.
* /doc/interpreter/module.mk: remove files from build system.