also cache parent_function in symbol_table scope
Inside a symbol table scope, we currently store a pointer to the
parent scope (if it exists), but this is not sufficient to prevent the
parent scope from being deleted.  For example, this can happen when a
function that contains an anonymous function definition is deleted.
If the anonymous function remains accessible, then resolving any
symbols that it references may fail when it is called.  Caching the
parent function avoids this problem as it increments the reference
count for the function, and that will prevent the scope object from
being deleted.

* symtab.cc (symbol_table::scope::set_parent): Also cache parent
function if it exists.

* symtab.h (symbol_table::scope::m_parent): New data member.
(symbol_table::scope::parent_fcn): New function.
(symbol_table::scope::dup): Also copy m_parent_fcn.