test: Mark certain bugs as xtest when run on Mac platforms (bug #52627).
* dlmread.cc: For BIST tests that fail due to bug #47413, protect test with
"testif ; ! ismac ()" so that it is not run on Mac platforms.
Duplicate failing test code, but protect it with:
%!xtest <47413>
%! ## Same test code as above, but intended only for test statistics on Mac.
%! if (! ismac ()), return; endif
so that it is run only on Mac platforms and contributes to xtest statistics.

* mappers.cc (Facos, Facosh, Fasin, Fasinh): Use run-time conditional
(! ismac * ()) to run regular BIST code on non-mac platforms.  Create xtest
which only executes on Macs to document failure.