configure.ac: Overhaul file (bug #53186).
* configure.ac: Disable OSMesa by default, but make --with-osmesa a configure
option (bug #53186).  Use "dnl" for comments in configure.ac which do not need
to be passed to the configure shell script.  Add more "dnl" and "##" comments
describing the configure process.  Re-order the configuration test macros to
make as much sense as is possible.  Define values through AC_DEFINE to 1,
rather than just to a defined state.  Use the phrase "Define to 1 ..." for
the comment field of any AC_DEFINE statements.  Rephrase some warning messages
for clarity and to get rid of the first person syntax.  Remove OSMesa from
build summary, change capitalization of OPENGL library to OpenGL in summary,
change position of ENABLE_FLOAT_TRUNCATE in summary.

* acinclude.m4: Define values through AC_DEFINE to 1, rather than just to a
defined state.  Use the phrase "Define to 1 ..." for the comment field of any
AC_DEFINE statements.