Improve absolute tolerance handling in ode-solvers (bug #54729).
* scripts/ode/private/AbsRel_norm.m: The absolute tolerance "AbsTol" is allowed
to be a vector, which is simply forwarded to this function.

In the odeset-structure the vector orientation is preserved, thus it makes only
sense to modify it's orientation when needed.

All other involved values "x" and "x_old" are ensured to be column-vectors,
therefore "max" yields a matrix, if "AbsTol" is a row-vector.

--HG--
branch : stable