Add -fillpage and -bestfit print options (bug #52838).
NEWS: Announce new options.

* __print_parse_opts__.m: New structure member resize_flag initialized to "".
Decode opts and set resize_flag to "fillpage" or "bestfit" if options found.
Validate that options are only used with a page-based format.
If resize option given, change the "paperposition" property of figure and
then let rest of print routine handle the work.
* __print_parse_opts__.m (gs_papersize): Add new second return argument which
is the "papersize" figure property measured in points.

* print.m: Update documentation with new options.

* contributors.in: Add Jun Wang to list of contributors.