Stop creating orphaned graphic handles if two "parent" properties given (bug #55322).
* graphics.cc (make_graphics_object): Increment loop index variable by 2 so
that property/value PAIRS are processed as one unit.  If "parent" property is
found, decrease loop index variable by 2 so that increment will leave index
in the same position in the list of property/value pairs.  This accounts
for the splice operation which removes two arguments from the list.

* test/bug-55322/bug-55322.tst: New test file for bug #55322.

* test/bug-55322/module.mk: Add test to build system.

--HG--
branch : stable