allow parsing of function handles of the form @foo.bar
This change is part of a fix for bug #51709.

* lex.ll (\@ | \@{S}*{FQIDENT}): Recognize function handles including
the name at the same time as "@" character that begins a function
handle or anonymous function so that we can recognize fully qualified
names.
(base_lexer::handle_identifier): Eliminate special handling of
function handle names since that is now done when recognizing the
"\@{S}*{FQIDENT}" pattern.

* oct-parse.yy (fcn_handle): Change rule to handle new meaning of
FCN_HANDLE token.