Fix stream positioning when using scanf with widths (bug #56200).
* oct-stream.cc (octave_scan (std::istream&, const scanf_format_elt&, T*)):
Rename variable "tmp" to "strbuf" for clarity.  Record position of input stream
"is" in variable "orig_pos" before any operations.  After scanning operation on
stringstream object "ss" check whether eof() is set (full width of characters
were used).  If not, restore "is" position back to value at start of function.
Find the number of characters read on stringstream object with tellg() and
then re-perform the extraction operation on input stream with the "correct"
width (number of characters).  Copy the failbit from the "ss" object to the
"is" object to propagate any pattern matching errors.