use Bison's push parser interface in the command line REPL
Using the push parser interface moves reading input and lexical
analysis outside of the parser.  This change shows how we might
also use this arrangement to allow the GUI terminal widget to be fully
in control of gathering user input.

A new configure option allows this change to be reverted so that
Octave will use the pull parser interface that has been used in all
previous versions of Octave.

* pt-eval.cc (tree_evaluator::repl): Use conditional compilation to
choose either push or pull parser.

* configure.ac (--disable-commnad-line-push-parser): New option.
Report option setting in summary message.