fix line count in lexer when mixing line and block comments (bug #57121)
* lex.ll (^{S}*{CCHAR}\{{S}*{NL}): Don't handle LINE_COMMENT_START
here since that can't happen.
(<LINE_COMMENT_START>^{S}*{CCHAR}\{{S}*{NL}): New pattern to match the
beginning of a block comment when we are already processing a series
of line-oriented comments.
(<LINE_COMMENT_START>{S}*{CCHAR}{ANY_EXCEPT_NL}*{NL}): Simplify.
There is no need to parse the comment text to attempt to match the
beginning of a block comment as that is now handled by a separate
pattern.