don't use singleton pattern for resource manager in GUI
Move ownership of the resource_manager object to the base_qobject
class.  Store reference to resource_manager object as a member
variable in most objects that need to access it.  Eliminate static
member functions in resource_managar class.  Provide a
get_resource_manager method in the base_qboject class and a
global __get_resource_manager__ function for those instances where we
can't easily provide local access to the resource_manager object.
There are many changes to many files, but most are straightforward.